% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_conf_quant_larger.R
\name{block_conf_quant_larger}
\alias{block_conf_quant_larger}
\title{block_conf_quant_larger}
\usage{
block_conf_quant_larger(
  Z,
  Y,
  block,
  method.list.all = NULL,
  opt.method = "Greedy",
  stat.null = NULL,
  switch = FALSE,
  null.max = 10^5,
  alpha = 0.1
)
}
\arguments{
\item{Z}{A numeric vector of treatment assignment. 0 indicates control and 1 treated.}

\item{Y}{A numeric vector contains response values of the group.}

\item{block}{A factor vector indicates to which blocks the units belong.}

\item{method.list.all}{A list that contains the method of the rank scores. It might vary among different blocks}

\item{opt.method}{Algorithm that is used for optimization. Available algorithms are "Greedy", "DP", "Mcknap", "LP", "ILP" and "LP_gurobi", "ILP_gurobi", "PWL_gurobi", "PWLint_gurobi". Gurobi installation is required for gurobi to be used.}

\item{stat.null}{Optional variable. You can enter the null distribution here by yourself. Otherwise it will be generated within the function.}

\item{switch}{Logical variable. If true, the function uses switching treatment and control trick when calculating p-value.}

\item{null.max}{The total amount of values we use to approximate the null distribution.}

\item{alpha}{scalar. It equals to 1-confidence of the confidence region}
}
\description{
A function to help determine confidence region with respect to k and c
}
