% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cad.R
\docType{data}
\name{cadmium}
\alias{cadmium}
\title{Cadmium level in blood after smoking}
\format{

}
\source{
NHANES, the US National Health and Nutrition Examination Survey, 2005-2006.
}
\usage{
data(cadmium)
}
\description{
Data from the 2005-2006 National Health and Nutrition Examination Survey to study how smoking influence the cadmium level in blood.
}
\keyword{datasets}
