% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_quant_scre.R
\name{ci_quant_scre}
\alias{ci_quant_scre}
\title{ci_quant_scre}
\usage{
ci_quant_scre(
  Z,
  Y,
  block,
  quantiles = NULL,
  alternative = "less",
  method.list.all = NULL,
  opt.method = "Greedy",
  ties = "fix",
  stat.null = NULL,
  switch = FALSE,
  null.max = 10^5,
  alpha = 0.1
)
}
\arguments{
\item{Z}{A numeric vector of treatment assignment. 0 indicates control and 1 treated.}

\item{Y}{A numeric vector contains response values of the group.}

\item{block}{A factor vector indicates to which blocks the units belong.}

\item{quantiles}{Specify which quantiles of treatment effects that you want to calculate confidence intervals. It should be a subvector of integers from 1 to number of units. If \code{quantiles = NULL}, it will calculate all quantiles. Default is NULL.}

\item{alternative}{Options: "lower", "upper", "two.sided", which will give lower one-sided, upper one-sided and two-sided confidence intervals, respectively. Default is "upper".}

\item{method.list.all}{A list that contains the type of the rank scores. It should be a list with length equal to the number of strata. Default is \code{NULL}.}

\item{opt.method}{Algorithm that is used for optimization. Available algorithms are "Greedy", "DP", "Mcknap", "LP", "ILP" and "LP_gurobi", "ILP_gurobi", "PWL_gurobi", "PWLint_gurobi". Gurobi installation is required for gurobi to be used. Default is "Greedy"}

\item{ties}{A subvector of \code{c("upper", "lower", "fix")} indicating which tie-dealing methods we use to calculate statistics. "upper" will use the method that will produce maximum statistic, while "lower" will get minimum. "fix" however will order the ties the same way as "first" method in rank function. Default is \code{c("upper", "lower", "fix")}.}

\item{stat.null}{Null distribution of test statistics. If no input, then it will be generated automatically.}

\item{switch}{Logical variable. If true, the function uses switching treatment and control label is the number of treated units is less than that of control units in one stratum. Default is "False".}

\item{null.max}{The total amount of values we use to approximate the null distribution. Default is $10^5$.}

\item{alpha}{scalar. Confidence level of the confidence region. Default is $0.1$.}
}
\value{
A list contains upper and lower limits of confidence intervals.
}
\description{
A function to help determine confidence region of treatment effects under SCRE.
}
\examples{
111

}
