% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_stat_block.R
\name{min_stat_block}
\alias{min_stat_block}
\title{min_stat_block}
\usage{
min_stat_block(
  Z,
  Y,
  block,
  k,
  c,
  method.list.all,
  opt.method = "Greedy",
  ties = c("upper", "lower", "random")
)
}
\arguments{
\item{Z}{A numeric vector of treatment assignment. 0 indicates control and 1 treated.}

\item{Y}{A numeric vector contains response values of the group.}

\item{block}{A factor vector indicates to which blocks the units belong.}

\item{k}{A parameter in hypothesis test that kth smallest treatment effect is less or equal to c}

\item{method.list.all}{A list that contains the method of the rank scores. If you use the same methods across all strata, method.list.all will only include 1 element stating that score. Otherwise, method.list.all[[s]] should be the method for stratum s. Available score functions: Wilcoxon, list(name = "Wilcoxon); Stephenson, list(name = "Stephenson", s = ?).}

\item{opt.method}{Algorithm that is used for optimization. Available algorithms are "Greedy", "DP", "Mcknap", "LP", "ILP" and "LP_gurobi", "ILP_gurobi", "PWL_gurobi", "PWLint_gurobi". Gurobi installation is required for gurobi to be used.}

\item{ties}{A subvector of c("upper", "lower", "random") indicating which tie-dealing methods we use to calculate statistics}
}
\description{
A function to calculate to calculate test statistics for null hypothesis: kth smallest treatment effect is less or equal to c.
}
