% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_trans.R
\name{opt_trans}
\alias{opt_trans}
\title{opt_trans}
\usage{
opt_trans(a)
}
\arguments{
\item{a}{Numeric vector that needs to be transformed.}
}
\description{
A function to get optimal monotone transformation
}
\keyword{Optimal}
\keyword{monotone}
\keyword{transformation}
