% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_block.R
\name{pval_block}
\alias{pval_block}
\title{pval_block}
\usage{
pval_block(
  Z,
  Y,
  block,
  k,
  c,
  method.list.all,
  Gam = 1,
  opt.method = "Greedy",
  ties = c("upper", "lower", "fix"),
  null.max = 10^5,
  stat.null = NULL,
  switch = FALSE
)
}
\arguments{
\item{Z}{A numeric vector of treatment assignment. 0 indicates control and 1 treated.}

\item{Y}{A numeric vector contains response values of the group.}

\item{block}{A factor vector indicates to which blocks the units belong.}

\item{k}{A parameter in hypothesis test that kth smallest treatment effect is less or equal to c}

\item{c}{A parameter in hypothesis test that kth smallest treatment effect is less or equal to c}

\item{method.list.all}{A list that contains the method of the rank scores. It might vary among different blocks}

\item{Gam}{A parameter for sensitivity analysis. If Gam = 1, the analysis is performed under normal setting. If Gam>1, then the analysis is under sensitivity setting.}

\item{opt.method}{Algorithm that is used for optimization. Available algorithms are "Greedy", "DP", "Mcknap", "LP", "ILP" and "LP_gurobi", "ILP_gurobi", "PWL_gurobi", "PWLint_gurobi". Gurobi installation is required for gurobi to be used.}

\item{ties}{A subvector of c("upper", "lower", "fix") indicating which tie-dealing methods we use to calculate p-values. "upper" will use the method that will produce maximum p-value, while "lower" will get minimum. "fix" however will order the ties the same way as "first" method in rank function.}

\item{null.max}{The total amount of values we use to approximate the null distribution.}

\item{stat.null}{Optional variable. You can enter the null distribution here by yourself. Otherwise it will be generated within the function.}

\item{switch}{Logical variable. If true, the function uses switching treatment and control trick when calculating p-value.}
}
\description{
A function to calculate valid p-value for null hypothesis: kth smallest treatment effect is less or equal to c.
}
