% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_block_sides.R
\name{pval_block_sides}
\alias{pval_block_sides}
\title{pval_block_sides}
\usage{
pval_block_sides(
  Z,
  Y,
  block,
  k,
  c,
  alternative = "greater",
  Gam = 1,
  method.list.all = NULL,
  null.max = 10^5
)
}
\arguments{
\item{Z}{A numeric vector of treatment assignment. 0 indicates control and 1 treated.}

\item{Y}{A numeric vector contains response values of the group.}

\item{block}{A factor vector indicates to which blocks the units belong.}

\item{k}{A parameter in hypothesis test}

\item{c}{A parameter in hypothesis test}

\item{alternative}{Options: "greater", "less", "two.sided", each defines differents alternative hypothesis}

\item{Gam}{A parameter for sensitivity analysis. If Gam = 1, the analysis is performed under normal setting. If Gam>1, then the analysis is under sensitivity setting.}

\item{method.list.all}{A list that contains the method of the rank scores. It might vary among different blocks}

\item{null.max}{The total amount of values we use to approximate the null distribution.}

\item{opt.method}{Algorithm that is used for optimization. Available algorithms are "Greedy", "DP", "Mcknap", "LP", "ILP" and "LP_gurobi", "ILP_gurobi", "PWL_gurobi", "PWLint_gurobi". Gurobi installation is required for gurobi to be used.}
}
\description{
A function to calculate valid p-value for generalized null hypothesis: two-sided or 2 one-sided tests
}
