% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_quantile_scre.R
\name{pval_quantile_scre}
\alias{pval_quantile_scre}
\title{pval_quantile_scre}
\usage{
pval_quantile_scre(
  Z,
  Y,
  block,
  k,
  c,
  alternative = "less",
  method.list.all = NULL,
  opt.method = "Greedy",
  ties = c("upper", "lower", "fix"),
  null.max = 10^5,
  switch = FALSE
)
}
\arguments{
\item{Z}{A numeric vector of treatment assignment. 0 indicates control and 1 treated.}

\item{Y}{A numeric vector contains response values of the group.}

\item{block}{A factor vector indicates to which blocks the units belong.}

\item{k}{A parameter in the null hypothesis.}

\item{c}{A parameter in the null hypothesis.}

\item{alternative}{Options: \code{"greater", "less", "two.sided"}, which correspond to different relationships between kth smallest treatment effect and c. For example, \code{alternative == "greater"} denotes null hypothesis that the kth smallest treatment effect is greater than c. Default is "less".}

\item{method.list.all}{A list that contains the type of the rank scores. It should be a list with length equal to the number of strata. Default is \code{NULL}.}

\item{opt.method}{Algorithm that is used for optimization. Available algorithms are \code{"Greedy", "DP", "Mcknap", "LP", "ILP" and "LP_gurobi", "ILP_gurobi", "PWL_gurobi", "PWLint_gurobi"}. Gurobi installation is required for gurobi to be used. Default is \code{"Greedy"}.}

\item{ties}{A subvector of \code{c("upper", "lower", "fix")} indicating which tie-dealing methods we use to calculate statistics. "upper" will use the method that will produce maximum statistic, while "lower" will get minimum. "fix" however will order the ties the same way as "first" method in rank function. Default is \code{c("upper", "lower", "fix")}.}

\item{null.max}{The total amount of values we use to approximate the null distribution. Default is \code{1e5}.}

\item{switch}{Logical variable. If true, the function uses switching treatment and control label is the number of treated units is less than that of control units in one stratum. Default is "False".}
}
\value{
A list contains upper and lower bound and a special value of $p$-values depending on the tie-dealing method.
}
\description{
A function to calculate valid p-value under SCRE for null hypothesis with respect to treatment effects. Specifically, we are interested in testing kth smallest treatment effect is less than, greater than or not equal to c.
}
\examples{
data("cadmium")
Y = cadmium$cadmium
block = cadmium$mset
Z = cadmium$z
### use stratified Wilcoxon test statistics for all strata. 
method.list.all = list()
method.list.all[[1]] = list(name = "Wilcoxon") 
### Test null hypothesis that 90\% quantile of treatment effects is less than or equal to 0.
p1 = pval_quantile_scre(Z=Z,Y=Y,block=block,k=floor(0.9*length(Y)),c=0, method.list.all=method.list.all,switch=TRUE)
### Test null hypothesis that 90\% quantile of treatment effects is greater than or equal to 1.
p2 = pval_quantile_scre(Z=Z,Y=Y,block=block,k=floor(0.9*length(Y)),c=1,alternative="greater", method.list.all=method.list.all,switch=TRUE)
}
