% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_quantile_sen.R
\name{pval_quantile_sen}
\alias{pval_quantile_sen}
\title{pval_quantile_sen}
\usage{
pval_quantile_sen(
  Z,
  Y,
  block,
  k,
  c,
  gam,
  method.list.all,
  opt.method = "Greedy",
  ties = c("upper", "lower", "fix"),
  switch = FALSE
)
}
\arguments{
\item{Z}{A numeric vector of treatment assignment. 0 indicates control and 1 treated.}

\item{Y}{A numeric vector contains response values of the group.}

\item{block}{A factor vector indicates to which blocks the units belong.}

\item{k}{A parameter in hypothesis test that at least k units have treatment effects less or equal to c}

\item{c}{A parameter in hypothesis test that at least k units have treatment effects less or equal to c}

\item{gam}{A parameter in sensitivity analysis controlling the odds of propensity scores.}

\item{method.list.all}{A list that contains the type of the rank scores. It should be a list with length equal to the number of strata. Default is \code{NULL}.}

\item{opt.method}{Algorithm that is used for optimization. Available algorithms are "Greedy", "DP", "Mcknap", "LP", "ILP" and "LP_gurobi", "ILP_gurobi", "PWL_gurobi", "PWLint_gurobi". Gurobi installation is required for gurobi to be used. Default is "Greedy"}

\item{ties}{A subvector of \code{c("upper", "lower", "fix")} indicating which tie-dealing methods we use to calculate statistics. "upper" will use the method that will produce maximum statistic, while "lower" will get minimum. "fix" however will order the ties the same way as "first" method in rank function. Default is \code{c("upper", "lower", "fix")}.}

\item{switch}{Logical variable. If true, the function uses switching treatment and control trick when calculating p-value. Default is "False".}
}
\value{
A list contains upper and lower bound and a special value of $p$-values depending on the tie-dealing method.
}
\description{
A function to calculate valid p-value under sensitivity analysis for null hypothesis with respect to treatment effects.
}
\examples{
111
 
}
