\name{QIoT-package}
\alias{QIoT-package}
\alias{QIoT}
\docType{package}
\title{
\packageTitle{QIoT}
}
\description{
\packageDescription{QIoT}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{QIoT}
\packageIndices{QIoT}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{QIoT}

Maintainer: \packageMaintainer{QIoT}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
