% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sen_block_conf_quant_larger.R
\name{sen_block_conf_quant_larger}
\alias{sen_block_conf_quant_larger}
\title{sen_block_conf_quant_larger}
\usage{
sen_block_conf_quant_larger(
  Z,
  Y,
  block,
  method.list.all = NULL,
  Gam,
  opt.method = "Greedy",
  ties = "fix",
  switch = FALSE,
  null.max = 10^5,
  alpha = 0.1
)
}
\arguments{
\item{Z}{A numeric vector of treatment assignment. 0 indicates control and 1 treated.}

\item{Y}{A numeric vector contains response values of the group.}

\item{block}{A factor vector indicates to which blocks the units belong.}

\item{method.list.all}{A list that contains the method of the rank scores. It might vary among different blocks}

\item{Gam}{A parameter for sensitivity analysis}

\item{opt.method}{Algorithm that is used for optimization. Available algorithms are "Greedy", "DP", "Mcknap", "LP", "ILP" and "LP_gurobi", "ILP_gurobi", "PWL_gurobi", "PWLint_gurobi". Gurobi installation is required for gurobi to be used.}

\item{ties}{ties = "upper", "lower" or "fix". It indicates which tie-dealing methods we use to calculate statistics.}

\item{switch}{Logical variable. If true, the function uses switching treatment and control trick when calculating p-value.}

\item{null.max}{The total amount of values we use to approximate the null distribution.}

\item{alpha}{scalar. It equals to 1-confidence of the confidence region}
}
\description{
A function to help determine confidence region with respect to k and c in sensitivity analysis
}
