% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_unemployed.R
\docType{data}
\name{unemployed}
\alias{unemployed}
\title{Austria unemployment change due to unemployment benefits.}
\format{
A data frame with 225821 rows and 175 variables.
}
\source{
Ruoqi Yu & Paul R. Rosenbaum (2022) Graded Matching for Large Observational Studies, Journal of Computational and Graphical Statistics, DOI: 10.1080/10618600.2022.2058001.
}
\usage{
data(unemployed)
}
\description{
This data concentrates on the reform in policies of unemployment benefits in Austria and how it influence the duration of unemployment.
}
\keyword{datasets}
