% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_quantile_scre.R
\name{pval_quantile_scre}
\alias{pval_quantile_scre}
\title{pval_quantile_scre}
\usage{
pval_quantile_scre(
  Z,
  Y,
  block,
  k,
  c,
  alternative = "less",
  method.list.all = NULL,
  opt.method = "Greedy",
  ties = c("upper", "lower", "fix"),
  null.max = 10^5
)
}
\arguments{
\item{Z}{A numeric vector of treatment assignment. 0 indicates control and 1 treated.}

\item{Y}{A numeric vector contains response values of the group.}

\item{block}{A factor vector indicates to which blocks the units belong.}

\item{k}{A parameter in the null hypothesis.}

\item{c}{A parameter in the null hypothesis.}

\item{alternative}{Options: "greater", "less", "two.sided", which correspond to null hypotheses in forms of $\tau_k\ge c$, $\tau_k\le c$, $\tau_k\ne c$, respectively. Default is "less".}

\item{method.list.all}{A list that contains the type of the rank scores. It should be a list with length equal to the number of strata. Default is \code{NULL}.}

\item{opt.method}{Algorithm that is used for optimization. Available algorithms are "Greedy", "DP", "Mcknap", "LP", "ILP" and "LP_gurobi", "ILP_gurobi", "PWL_gurobi", "PWLint_gurobi". Gurobi installation is required for gurobi to be used. Default is "Greedy"}

\item{ties}{A subvector of \code{c("upper", "lower", "fix")} indicating which tie-dealing methods we use to calculate statistics. "upper" will use the method that will produce maximum statistic, while "lower" will get minimum. "fix" however will order the ties the same way as "first" method in rank function. Default is \code{c("upper", "lower", "fix")}.}

\item{null.max}{The total amount of values we use to approximate the null distribution. Default is $10^5$.}
}
\value{
A list contains upper and lower bound and a special value of $p$-values depending on the tie-dealing method.
}
\description{
A function to calculate valid p-value under SCRE for null hypothesis with respect to treatment effects.
}
\examples{
data("cadmium")
Y = cadmium$cadmium
block = cadmium$mset
Z = cadmium$z
method.list.all = list()
method.list.all[[1]] = list(name = "Wilcoxon") 
pval_quantile_scre(Z=Z,Y=Y,block=block,k=floor(0.8*length(Y)),c=0,method.list.all=method.list.all)
}
