% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_quantile_sen.R
\name{pval_quantile_sen}
\alias{pval_quantile_sen}
\title{pval_quantile_sen}
\usage{
pval_quantile_sen(
  Z,
  Y,
  block,
  k,
  c,
  alternative = "less",
  gam = 1,
  method.list.all = NULL,
  opt.method = "Greedy",
  ties = c("upper", "lower", "fix")
)
}
\arguments{
\item{Z}{A numeric vector of treatment assignment. 0 indicates control and 1 treated.}

\item{Y}{A numeric vector contains response values of the group.}

\item{block}{A factor vector indicates to which blocks the units belong.}

\item{k}{A parameter in the null hypothesis.}

\item{c}{A parameter in the null hypothesis.}

\item{alternative}{Options: \code{"greater", "less", "two.sided"}, which correspond to different relationships between kth smallest treatment effect and c. For example, \code{alternative == "greater"} denotes null hypothesis that the kth smallest treatment effect is greater than c. Default is "less".}

\item{gam}{A parameter in sensitivity analysis controlling the odds of propensity scores. Default is 1.}

\item{method.list.all}{A list that contains the type of the rank scores. It should be a list with length equal to the number of strata. Default is \code{NULL}.}

\item{opt.method}{Algorithm that is used for optimization. Available algorithms are \code{"Greedy", "DP", "Mcknap", "LP", "ILP" and "LP_gurobi", "ILP_gurobi", "PWL_gurobi", "PWLint_gurobi"}. Gurobi installation is required for gurobi to be used. Default is \code{"Greedy"}.}

\item{ties}{A subvector of \code{c("upper", "lower", "fix")} indicating which tie-dealing methods we use to calculate statistics. "upper" will use the method that will produce maximum statistic, while "lower" will get minimum. "fix" however will order the ties the same way as "first" method in rank function. Default is \code{c("upper", "lower", "fix")}.}
}
\value{
A list contains upper and lower bound and a special value of $p$-values depending on the tie-dealing method.
}
\description{
A function to calculate valid p-value under sensitivity analysis for null hypothesis with respect to treatment effects. Specifically, we are interested in testing kth smallest treatment effect is less than, greater than or not equal to c.
}
\examples{
111
 
}
