% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sen_ls_p.R
\name{sen_ls_p}
\alias{sen_ls_p}
\title{sen_ls_p}
\usage{
sen_ls_p(
  Z,
  Y,
  block,
  k,
  c,
  gam,
  method.list.all,
  opt.method = "Greedy",
  switch = FALSE
)
}
\arguments{
\item{Z}{A numeric vector of treatment assignment. 0 indicates control and 1 treated.}

\item{Y}{A numeric vector contains response values of the group.}

\item{block}{A factor vector indicates to which blocks the units belong.}

\item{k}{A parameter in hypothesis test that at least k units have treatment effects less or equal to c}

\item{c}{A parameter in hypothesis test that at least k units have treatment effects less or equal to c}

\item{gam}{A parameter for sensitivity analysis}

\item{method.list.all}{A list that contains the method of the rank scores. It might vary among different blocks}

\item{opt.method}{Algorithm that is used for optimization. Available algorithms are "Greedy", "DP", "Mcknap, "LP", "ILP".}

\item{switch}{Logical variable. If true, the function uses switching treatment and control trick when calculating p-value.}
}
\description{
A function to calculate valid p-value for null hypothesis: kth smallest treatment effect is less or equal to c.
}
